+++ src/main/java/com/snake/game/SquarePanel.java
@@ -1,0 +1,19 @@
+package com.snake.game;
+
+import java.awt.Color;
+import javax.swing.JPanel;
+
+public class SquarePanel extends JPanel{
+
+    private static final long serialVersionUID = 1L;
+
+    public SquarePanel(Color d){
+        this.setBackground(d);
+    }
+
+    public void ChangeColor(Color d){
+        this.setBackground(d);
+        this.repaint();
+    }
+
+}
